requirements.txt
django
cryptography
requests
djangorestframework
django-cors-headers